backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.41
alembic==1.16.4
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
paho-mqtt==1.6.1
jsonschema==4.24.0
pandas==2.2.3
